Rename `addons` --> `hub-addon`
