+++ pkg/cmd/install/hubaddon/scenario/addon/appmgr/clusterrole.yaml
