+++ pkg/cmd/install/hubaddon/scenario/addon/appmgr/clusterrole_agent.yaml
