+++ pkg/cmd/install/hubaddon/scenario/addon/appmgr/clusterrole_binding.yaml
