+++ pkg/cmd/install/hubaddon/scenario/addon/appmgr/crd_channel.yaml
