+++ pkg/cmd/install/hubaddon/scenario/addon/appmgr/crd_subscription.yaml
