+++ pkg/cmd/install/hubaddon/scenario/addon/appmgr/crd_subscriptionstatuses.yaml
