+++ pkg/cmd/install/hubaddon/scenario/addon/appmgr/deployment_appsubsummary.yaml
