+++ pkg/cmd/install/hubaddon/scenario/addon/appmgr/deployment_placementrule.yaml
