+++ pkg/cmd/install/hubaddon/scenario/addon/appmgr/service.yaml
