+++ pkg/cmd/install/hubaddon/scenario/addon/appmgr/service_account.yaml
