+++ pkg/cmd/install/hubaddon/scenario/resources.go
