+++ test/integration-test.mk
@@ -30,5 +30,5 @@ clean-integration-test:
 clean: clean-integration-test
 
 test-integration: ensure-kubebuilder-tools
-	go test -v ./pkg/cmd/addon/enable  ./pkg/cmd/addon/disable  ./pkg/cmd/install/addons 
+	go test -v ./pkg/cmd/addon/enable  ./pkg/cmd/addon/disable  ./pkg/cmd/install/hubaddon 
 .PHONY: test-integration
